requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data processing
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
httpx==0.25.2

# Task queue and caching
celery==5.3.4
redis==5.0.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data validation and utilities
validators==0.22.0
python-slugify==8.0.1
geopy==2.4.0

# Rate limiting and retries
tenacity==8.2.3
ratelimit==2.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and monitoring
structlog==23.2.0